Refactor: 브랜치 통합을 위한 코드 구조 정리
