+++ .gitignore
@@ -1,16 +1,17 @@
-# Java compiled class files
+# =================
+# Gradle / Build
+# =================
+.gradle/
+build/
+gradlew
+gradlew.bat
+gradle-wrapper.jar
+gradle-wrapper.properties
+
+# =================
+# Java / Kotlin
+# =================
 *.class
-
-# Log files
-*.log
-
-# BlueJ files
-*.ctxt
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files
 *.jar
 *.war
 *.nar
@@ -19,29 +20,11 @@
 *.tar.gz
 *.rar
 
-# Virtual machine crash logs
-hs_err_pid*
-replay_pid*
-
-# Kotlin Gradle plugin data
-.kotlin/
-
-# Android Studio / IntelliJ specific
+# =================
+# Android Studio / IntelliJ
+# =================
 *.iml
 .idea/
-.gradle/
-local.properties
-
-# Mac OS
-.DS_Store
-
-# Build outputs
-/build/
-captures/
-.externalNativeBuild/
-.cxx/
-
-# Android specific IDE cache
 /.idea/caches/
 /.idea/libraries/
 /.idea/modules.xml
@@ -49,59 +32,50 @@ captures/
 /.idea/navEditor.xml
 /.idea/assetWizardSettings.xml
 
-
-##########################
-
-
-# Google API Keys 및 Firebase Secrets (필수)
+# =================
+# Firebase / Google Services
+# =================
 /app/google-services.json
-
-# Android Studio와 Gradle 자동 생성 파일
-.gradle/
-/local.properties
-/build
-
-# Firebase CLI 및 Functions 관련
+google-services.json
 .firebase/
 .firebaserc
-node_modules/local.properties
+firebase.json
+firestore.rules
+storage.rules
+firestore.indexes.json
+database.rules.json
+firebase.debug.log
+/backup-data/
+
+# =================
+# Local properties
+# =================
 local.properties
+/local.properties
+node_modules/local.properties
 
+# =================
+# Build outputs
+# =================
+/build/
+captures/
+.externalNativeBuild/
+.cxx/
 
-    gitignore
-    # =================
-    # Gradle / Build
-    # =================
-    .gradle/
-    build/
+# =================
+# Logs
+# =================
+*.log
+hs_err_pid*
+replay_pid*
 
-    # =================
-    # Firebase / Google Services
-    # =================
-    # Firebase Emulator의 백업 데이터
-    /backup-data/
+# =================
+# OS
+# =================
+.DS_Store
+.mtj.tmp/
 
-    # Firebase 로컬 설정 파일
-    .firebase/
-    firebase.json
-    firestore.rules
-    storage.rules
-    firestore.indexes.json
-    database.rules.json
-    firebase.debug.log
-    *.log
-
-    # Google Services 설정 파일 (로컬에서만 사용)
-    google-services.json
-
-    # =================
-    # Local properties
-    # =================
-    # API 키와 같은 비밀 정보 저장
-    local.properties
-
-    # =================
-    # IDE
-    # =================
-    .idea/
-    *.iml
+# =================
+# Kotlin
+# =================
+.kotlin/
