+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,10 @@
 
         <activity
             android:name=".MainActivity"
+            android:exported="true" />
+
+        <activity
+            android:name=".map.MidpointActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
