+++ app/src/main/java/com/moyeoyo/app/data/model/Group.kt
@@ -1,17 +1,17 @@
 package com.moyeoyo.app.data.model
 
+import com.google.firebase.firestore.DocumentId
 import com.google.firebase.Timestamp
 
-// groups/{groupId} 문서 모델
 data class Group(
-    val groupId: String = "",
+    @DocumentId
+    val id: String = "",
     val groupName: String = "",
     val hostUid: String = "",
     val memberUids: List<String> = emptyList(),
-    val confirmedPlace: PlaceCandidate? = null,
+    val confirmedPlace: Map<String, Any>? = null,
     val confirmedTime: Timestamp? = null,
-    val status: String = "VOTING",
-    val createdAt: Timestamp? = null
+    val status: String = "VOTING"
 )
